Allow the driver to operate with a 16k page size kernel
Raspberry Pi OS defaults to a 16kb page size kernel. The driver
seemingly works if we operate as if we're on a 4k page kernel in that
case, though I added a warning in this configuration, since it's
probably safer for the user to just switch to a 4k page size kernel, as
that path would be better tested.

Co-authored-by: Janne Grunau <j@jannau.net>